% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bhai_create_bargraph}
\alias{bhai_create_bargraph}
\title{creats bar graph}
\usage{
bhai_create_bargraph(
  bhai_use_this_data,
  bhai_use_this_column,
  bhai_use_this_title_for_xAxis = "supply title for x-axis",
  bhai_use_this_title_for_yAxis = "supply title for y-axis"
)
}
\arguments{
\item{bhai_use_this_data}{pass data here}

\item{bhai_use_this_column}{pass column}

\item{bhai_use_this_title_for_xAxis}{title for x axis}

\item{bhai_use_this_title_for_yAxis}{title for y axis}
}
\value{
returns graph
}
\description{
creats bar graph
}
\examples{
 bhai_create_bargraph(mtcars,cyl,"cyl","count")
}
