% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bhai_create_boxplot}
\alias{bhai_create_boxplot}
\title{created boxplot}
\usage{
bhai_create_boxplot(
  bhai_use_this_data,
  xData,
  yData,
  bhai_use_this_category_for_colour
)
}
\arguments{
\item{bhai_use_this_data}{pass the data}

\item{xData}{data on x axis}

\item{yData}{data on y axis}

\item{bhai_use_this_category_for_colour}{category for colour}
}
\description{
created boxplot
}
