% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bhai_create_histogram}
\alias{bhai_create_histogram}
\title{creates histogram}
\usage{
bhai_create_histogram(
  bhai_use_this_data,
  bhai_use_this_column,
  Bhai_use_this_colour = "#69b3a2",
  Bhai_use_this_binwidth = 5
)
}
\arguments{
\item{bhai_use_this_data}{pass data}

\item{bhai_use_this_column}{pass column to be used here}

\item{Bhai_use_this_colour}{pass the colour}

\item{Bhai_use_this_binwidth}{pass the bin-width here}
}
\value{
it returns the graph drawn
}
\description{
creates histogram
}
\examples{
bhai_create_histogram(bhai_use_this_data=cars,
bhai_use_this_column=dist,
Bhai_use_this_colour='red',
Bhai_use_this_binwidth = 3 )

}
