% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bhai_create_scatterplot}
\alias{bhai_create_scatterplot}
\title{creates scatterplot}
\usage{
bhai_create_scatterplot(
  bhai_use_this_data,
  xData,
  yData,
  bhai_use_this_category_for_colour
)
}
\arguments{
\item{bhai_use_this_data}{pass the data frame}

\item{xData}{data on x axis}

\item{yData}{data on y axis}

\item{bhai_use_this_category_for_colour}{pass the colour}
}
\description{
creates scatterplot
}
