% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{kanka_create_boxplot}
\alias{kanka_create_boxplot}
\title{oluşturmak Boxplot}
\usage{
kanka_create_boxplot(
  kanka_use_this_data,
  xData,
  yData,
  kanka_use_this_category_for_renk
)
}
\arguments{
\item{kanka_use_this_data}{pass data}

\item{xData}{data on x-axis}

\item{yData}{data on y-axis}

\item{kanka_use_this_category_for_renk}{pass category}
}
\value{
returns graph
}
\description{
oluşturmak Boxplot
}
\examples{
kanka_create_boxplot(iris,
xData =  Species,
yData = Sepal.Length,
Species)
}
