% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{kanka_create_histogram}
\alias{kanka_create_histogram}
\title{oluşturmak Histogram}
\usage{
kanka_create_histogram(
  kanka_use_this_data,
  kanka_use_this_column,
  kanka_use_this_renk = "#69b3a2",
  kanka_use_this_binwidth = 5
)
}
\arguments{
\item{kanka_use_this_data}{pass data}

\item{kanka_use_this_column}{pass column}

\item{kanka_use_this_renk}{pass renk}

\item{kanka_use_this_binwidth}{pass binwidth}
}
\value{
returns graph
}
\description{
oluşturmak Histogram
}
\examples{
kanka_create_histogram (cars,
dist,
'red',
3 )
}
